chore: code review cleanup
